Change Amount Type to Double

+++ ProductsAPI/Models/Production.cs
@@ -5,12 +5,12 @@ public class Production
     public Guid ProductionId { get; set; }
     public int Code { get; set; }
     public int ProductCode { get; set; }
-    public int Amount { get; set; }
+    public double Amount { get; set; }
     public DateTime FabricationDate { get; set; }
     public int Validity { get; set; }
     public double RecipesQuantity { get; set; }
 
-    public Production(Guid productionId, int code, int productCode, int amount, DateTime fabricationDate, int validity, double recipesQuantity)
+    public Production(Guid productionId, int code, int productCode, double amount, DateTime fabricationDate, int validity, double recipesQuantity)
     {
         ProductionId = productionId;
         Code = code;
